+++ playground/app/composables/usePageMeta.ts
@@ -54,7 +54,7 @@ const components: IComponentInfo[] = [
   { id: 'calendar', description: '(~) A calendar tool for choosing individual dates, multiple dates, or date spans.', iconData: Calendar1Icon },
   { id: 'checkbox', description: '(-) An input element to toggle between checked and unchecked states.', iconData: FormIcon },
   { id: 'collapsible', description: '(+) A collapsible component for showing or hiding its content.', iconData: ItemIcon, iconUi: 'pr-px pb-[3px]' },
-  { id: 'chip', description: '(-) An indicator of a numeric value or a state.', iconData: ItemIcon, iconUi: 'pr-px pb-[3px]' },
+  { id: 'chip', description: '(+) An indicator of a numeric value or a state.', iconData: ItemIcon, iconUi: 'pr-px pb-[3px]' },
   { id: 'confetti', description: '(~) Performant confetti animation in the browser.', iconData: BoldSparkleIcon },
   { id: 'container', description: '(~) A container lets you center and constrain the width of your content.', iconData: GroupIcon },
   { id: 'countdown', description: '(~) Countdown with options control.', iconData: Clock1Icon },