+++ src/runtime/utils/tv.ts
@@ -13,10 +13,24 @@ const twMergeConfig = {
       'context-edge-light', 'context-edge-dark'
     ],
     'b24-colors': [
-      'air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot',
-      'air-secondary', 'air-secondary-alert', 'air-secondary-accent', 'air-secondary-accent-1', 'air-secondary-accent-2', 'air-secondary-no-accent',
-      'air-tertiary', 'air-tertiary-accent', 'air-tertiary-no-accent',
+      // primary ////
+      'air-primary', 'air-primary-inverted', // ---- ////
+      'air-primary-success', 'air-primary-success-inverted',
+      'air-primary-alert', 'air-primary-alert-inverted',
+      'air-primary-copilot',
+      'air-primary-warning',
+      'air-primary-no-accent', 'air-primary-no-accent-inverted',
+      // secondary ////
+      'air-secondary',
+      'air-secondary-alert',
+      'air-secondary-accent', 'air-secondary-accent-1', 'air-secondary-accent-2',
+      'air-secondary-no-accent', 'air-secondary-no-accent-1',
+      // tertiary ////
+      'air-tertiary',
+      'air-tertiary-accent', 'air-tertiary-no-accent',
+      // custom ////
       'air-selection', 'air-boost',
+      // old ////
       'old-style-default', 'old-style-danger', 'old-style-success',
       'old-style-warning', 'old-style-primary', 'old-style-secondary',
       'old-style-collab', 'old-style-ai'