+++ src/theme/chip.ts
@@ -3,6 +3,7 @@
  * As a source we will choose `counter` and `labels`
  * ---
  * @link /api_d7/bitrix/ui/labels/index.php
+ * @see src/runtime/air-design-tokens/components/badge-counter.css
  * @see bitrix/js/ui/cnt/src/cnt-color.js
  * @see bitrix/js/ui/cnt/ui.cnt.css
  * @see bitrix/js/ui/label/src/label-color.js
@@ -13,29 +14,58 @@
 export default {
   slots: {
     root: [
-      '',
-      'relative inline-flex items-center justify-center shrink-0'
+      'relative shrink-0',
+      'inline-flex items-center justify-center'
     ].join(' '),
     base: [
-      'context-light',
+      'ui-counter__scope --air',
+      'font-[family-name:var(--ui-font-family-primary)]',
+      'font-(--ui-font-weight-medium)',
       'select-none',
-      'rounded-sm',
-      'flex items-center justify-center',
-      'p-1',
-      'bg-(--b24ui-background-hover)',
-      'text-(--b24ui-color)',
-      'leading-none font-semibold font-b24-secondary',
-      'whitespace-nowrap'
+      'relative',
+      'min-w-(--ui-counter-size) h-(--ui-counter-size)',
+      'py-0 px-(--ui-counter-inline-space)',
+      'inline-flex items-center justify-center',
+      'bg-(--b24ui-background)',
+      'rounded-(--ui-counter-current-size)',
+      'ring-(length:--b24ui-border-width) ring-(--b24ui-border-color)',
+      'text-center align-middle',
+      'text-(length:--ui-counter-font-size) text-(--b24ui-color)',
+      'leading-[inherit]',
+      'overflow-hidden',
+      'z-1'
+      // fix
+      // 'rounded-sm',
+      // 'flex items-center justify-center',
+      // 'p-1',
+      // 'bg-(--b24ui-background-hover)',
+      // 'text-(--b24ui-color)',
+      // 'text-(length:--ui-counter-font-size)',
+      // 'whitespace-nowrap'
+    ].join(' '),
+    trailingIcon: [
+      'size-(--ui-counter-size)',
+      'text-inherit',
+      'text-(length:--ui-counter-symbol-font-size)',
+      'opacity-96',
+      'tracking-(--ui-letter-spacing-xl)',
+      'me-(--ui-counter-symbol-compensation) empty:me-[0]'
     ].join(' ')
   },
   variants: {
     color: {
       'air-primary': { base: 'air-primary' },
+      'air-primary-inverted': { base: 'air-primary-inverted' },
       'air-primary-success': { base: 'air-primary-success' },
+      'air-primary-success-inverted': { base: 'air-primary-success-inverted' },
       'air-primary-alert': { base: 'air-primary-alert' },
+      'air-primary-alert-inverted': { base: 'air-primary-alert-inverted' },
+      'air-primary-warning': { base: 'air-primary-warning' },
       'air-primary-copilot': { base: 'air-primary-copilot' },
-      'air-secondary': { base: 'air-secondary' },
-      'air-secondary-alert': { base: 'air-secondary-alert' },
+      'air-secondary': { base: 'air-secondary-no-accent-1' },
+      'air-secondary-accent': { base: 'air-primary-no-accent' },
+      'air-secondary-accent-1': { base: 'air-primary-no-accent-inverted' },
+      'air-secondary-no-accent': { base: 'air-secondary-no-accent' },
       // @deprecate ////
       'default': { base: 'old-style-default' },
       'danger': { base: 'old-style-danger' },
@@ -45,26 +75,11 @@ export default {
       'secondary': { base: 'old-style-secondary' },
       'collab': { base: 'old-style-collab' },
       'ai': { base: 'old-style-ai' }
-      // default: 'bg-base-500 dark:bg-base-900 dark:text-base-150',
-      // danger: 'bg-red-500 dark:bg-red-600 dark:text-red-250',
-      // success: 'bg-green-500 dark:bg-green-600 dark:text-green-250',
-      // warning: 'bg-orange-500 dark:bg-orange-600 dark:text-orange-250',
-      // primary: 'bg-blue-500 dark:bg-blue-600 dark:text-blue-250',
-      // secondary: 'bg-cyan-350 dark:bg-cyan-500',
-      // collab: 'bg-collab-500 dark:bg-collab-800 dark:text-collab-250',
-      // ai: 'bg-ai-500 dark:bg-ai-800 dark:text-ai-250',
-      // link: 'bg-base-900/85 text-white dark:bg-white/85 dark:text-base-900'
     },
     size: {
-      '3xs': 'h-1 min-w-1 text-[0px] text-transparent p-0',
-      '2xs': 'h-1.5 min-w-1.5 text-[0px] text-transparent p-0',
-      'xs': 'h-xs2 min-w-xs2 text-[0px] text-transparent  p-0',
-      'sm': 'h-sm min-w-sm text-4xs font-regular',
-      'md': 'h-md min-w-md text-3xs',
-      'lg': 'h-lg min-w-lg rounded-md text-xs',
-      'xl': 'h-lg min-w-lg rounded-md text-xs',
-      '2xl': 'h-lg min-w-lg rounded-md text-xs',
-      '3xl': 'h-lg min-w-lg rounded-md text-xs'
+      'sm': 'ui-counter-sm font-(--ui-font-weight-regular)',
+      'md': 'ui-counter-md',
+      'lg': 'ui-counter-lg'
     },
     position: {
       'top-right': 'top-0 right-0',
@@ -77,7 +92,17 @@ export default {
     },
     standalone: {
       true: '',
-      false: 'ring ring-(--b24ui-border-color) absolute'
+      false: 'absolute'
+    },
+    hideZero: {
+      true: {
+        base: 'data-[value=0]:hidden'
+      }
+    },
+    oneDigit: {
+      true: {
+        base: 'px-0'
+      }
     }
   },
   compoundVariants: [